+++ apps/sim/lib/table/index.ts
@@ -18,3 +18,4 @@ export * from './service'
 export * from './sql'
 export * from './types'
 export * from './validation'
+export * from './workflow-groups/service'
